Add products API
Added products api that can grab products. Used the information gathered from there to grab 6 products. Only issue is the products are of the same brand and company. Used those six products to populate the seeds folder for products. I will no begin working on the models for the products. Eventually, I will diversify the products to include more brands and products